Created class for loading tile types. Dont want to load tile types everytime a new map is created.
