+++ src/main/java/mario/Player.java
@@ -3,6 +3,7 @@ package mario;
 import mario.tiles.FloorTile;
 import mario.tiles.KeyTile;
 import mario.tiles.LavaTile;
+import mario.tiles.MarioTile;
 import nl.han.ica.oopg.collision.CollidedTile;
 import nl.han.ica.oopg.collision.CollisionSide;
 import nl.han.ica.oopg.collision.ICollidableWithGameObjects;
@@ -25,6 +26,7 @@ public class Player extends AnimatedSpriteObject implements ICollidableWithTiles
     private int movementSpeed = 5;
 
     public Player(MainApp app) {
+
         super(new Sprite(MainApp.MEDIA_URL.concat("sprites/characters/mario.png")), 7);
         this.app = app;
         this.jumpSound = new Sound(app, MainApp.MEDIA_URL.concat(("sounds/jump_11.wav")));
@@ -35,6 +37,7 @@ public class Player extends AnimatedSpriteObject implements ICollidableWithTiles
     }
 
     private void initPlayer() {
+
         this.setSpeed(this.movementSpeed);
         this.setGravity(this.gravity);
     }
@@ -43,6 +46,7 @@ public class Player extends AnimatedSpriteObject implements ICollidableWithTiles
     public void keyPressed(int keyCode, char key) {
 
         if (keyCode == app.RIGHT) {
+            nextFrame();
             setDirectionSpeed(90, this.movementSpeed);
         } else if (keyCode == app.LEFT) {
             setDirectionSpeed(270, this.movementSpeed);
@@ -52,19 +56,19 @@ public class Player extends AnimatedSpriteObject implements ICollidableWithTiles
             this.jumpSound.play();
         }
     }
-
+    /*
     @Override
     public void keyReleased(int keyCode, char key) {
         this.setSpeed(0);
 
-        /*if (keyCode == app.RIGHT) {
+        if (keyCode == app.RIGHT) {
             setDirectionSpeed(0, 0);
         } else if (keyCode == app.LEFT) {
             setDirectionSpeed(0, 0);
         } else if (key == ' ') {
             setDirectionSpeed(0, 0);
-        }*/
-    }
+        }
+    }*/
 
     @Override
     public void gameObjectCollisionOccurred(List<GameObject> list) {
@@ -106,6 +110,7 @@ public class Player extends AnimatedSpriteObject implements ICollidableWithTiles
             
             if (tile.getTile() instanceof KeyTile) {
                 try {
+
                     this.app.getTileMap().setTile((int) tileIndexLocation.x, (int) tileIndexLocation.y, -1);
                 } catch (TileNotFoundException e) {
                     e.printStackTrace();
