+++ src/main/java/mario/tiles/MarioTile.java
@@ -3,7 +3,7 @@ package mario.tiles;
 import nl.han.ica.oopg.objects.Sprite;
 import nl.han.ica.oopg.tile.Tile;
 
-public class MarioTile extends Tile {
+public abstract class MarioTile extends Tile {
 
     private static final int tileSize = 64;
